// src/domain/enums/service-type.enum.ts
export enum ServiceType {
  // Serviços de Cuidadores
  APLICACAO_INJECAO = 'Aplicação de Injeção',
  CURATIVO = 'Realização de Curativo',
  AFERICAO_PRESSAO = 'Aferição de Pressão',

  // Serviços de Massagista
  MASSAGEM_RELAXANTE = 'Massagem Relaxante',
  DRENAGEM_LINFATICA = 'Drenagem Linfática',
  SHIATSU = 'Shiatsu',

  // Serviços de Fisioterapia
  FISIOTERAPIA_MOTORA = 'Fisioterapia Motora',
  FISIOTERAPIA_RESPIRATORIA = 'Fisioterapia Respiratória',
  PILATES_CLINICO = 'Pilates Clínico',
  
  // Serviços de Psicologia
  SESSAO_TERAPIA_INDIVIDUAL = 'Sessão de Terapia Individual',
  TERAPIA_CASAL = 'Terapia de Casal',
  AVALIACAO_PSICOLOGICA = 'Avaliação Psicológica',
  
  // Serviços de Médico
  CONSULTA_CLINICA_GERAL = 'Consulta Clínica Geral',
  EMISSAO_ATESTADO = 'Emissão de Atestado',
  PRESCRICAO_MEDICAMENTOS = 'Prescrição de Medicamentos',

  TRANSPORTE_PACIENTE = 'Transporte de Paciente',
  ACOMPANHAMENTO_CONSULTA = 'Acompanhamento em Consulta',
  SUPORTE_ADMINISTRATIVO = 'Suporte Administrativo',

}